refactor(ast): improve parser error handling
